requirements.txt
django==1.10.5
pytz==2016.6.1
psycopg2==2.6.2
django-authtools==1.5.0
